fix(ui): update library and clarify settings UX
- Updated `hyperisland-kit` to 0.3.3 to resolve Issue #9 (Shade visibility bug).
- Redesigned `IslandSettingsControl`:
  - Updated text descriptions to clearly explain 'Float' (Heads-up) vs 'Shade' (System Panel) behavior.
  - Improved Timeout Slider: Now displays seconds (0-10s) clearly in the header and uses full width for easier adjustment.
- Updated English and Spanish translations.

Closes #9

+++ app/src/main/java/com/d4viddf/hyperbridge/ui/components/IslandSettingsControl.kt
@@ -96,7 +96,7 @@ fun IslandSettingsControl(
                 value = timeoutSec,
                 onValueChange = { seconds ->
                     // FIX: Pass seconds directly as Long (No * 1000 conversion)
-                    onUpdate(config.copy(timeout = seconds.toInt()))
+                    onUpdate(config.copy(timeout = seconds.toLong()))
                 },
                 valueRange = 0f..10f,
                 steps = 9,
